Messing about to change the number of images handled
Remove fine name extension so outputs are .png
Handle a number of images that does not fit entirely into the
numCLBuffers. So the last loop over them only uses as many buffers as
needed.
On exit, only free the buffers that were used on the last loop.

Added timings for the file IO. It gets stored for every thread so it can
be subtracted per image from the total to get closer to the real per
image overhead for just the OpenCL side of things.

(Indentation got screwed by emacs)